Converting ember-test-types-task to addResult
